backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pandas>=2.0.0
sqlalchemy>=2.0.23
pydantic>=2.0.0
pydantic-settings>=2.1.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.6
redis>=5.0.0
celery>=5.3.0
openpyxl>=3.1.2
pytest>=7.4.3
pytest-asyncio>=0.21.1
requests>=2.31.0
python-dotenv>=1.0.0
supabase
python-jose[cryptography]
simpleeval

# Production dependencies
slowapi>=0.1.9
structlog>=24.1.0
prometheus-client>=0.20.0
sentry-sdk[fastapi]>=1.40.0
ruff>=0.1.5
mypy>=1.8.0
